+++ R/data-gridlayer-basic.R
@@ -1,3 +1,4 @@
+setClassUnion("oneDim", c("vector", "table", "array"))
 
 #gridlayer basic class
 	#class definition
@@ -9,7 +10,7 @@
 			tessellation ="numeric",
 			gridclass = "character",
 			names = "character",
-			values= "vector",
+			values= "oneDim",
 			length= "integer"
 		)
 	
@@ -20,7 +21,7 @@ setMethod("show", signature ="gridlayer",
 		#	cat(paste(class(object), "of", object@grid ,"with", object@length, class(object@values), "values\n", sep=" "))
 		#	cat(object@values, fill=TRUE)
 			
-			actGrid<-get(object@grid)
+			actGrid<-dynGet(object@grid, minframe=0L)
 			
 			
 			cat(paste("class        : ", class(object),"\n", sep=""))